+++ src/lib/validators.ts
@@ -243,6 +243,13 @@ export namespace Validators {
         return num;
     }
 
+    export function url(errorMessage = 'Invalid URL') {
+        return stringValidator(inputValue => {
+            const pattern = /^((([A-Za-z]{3,9}:(?:\/\/)?)(?:[-;:&=+$,\w]+@)?[A-Za-z0-9.-]+(:[0-9]+)?|(?:www.|[-;:&=+$,\w]+@)[A-Za-z0-9.-]+)((?:\/[+~%\/.\w-_]*)?\??[-+=&;%@.\w_]*#?[\w]*)?)$/;
+            return pattern.test(inputValue) || errorMessage;
+        });
+    }
+
     export function createCallbackList(callbacks: CallbackDefinition<any>[]) {
         return callbacks.map(callback => createCallbackByKey(callback.action, callback.arguments));
     }