+++ src/main.ts
@@ -7,10 +7,10 @@ import FormVue from './FormVue.vue';
 
 // OPTIONAL ELEMENTS
 import BaseInput from './components/fields/base_input.vue';
-import OnTextfieldText from "./components/fields/textfield_text.vue";
-import OnTextfieldEmail from "./components/fields/textfield_email.vue";
-import OnTextfieldNumber from "./components/fields/textfield_number.vue";
-import OnTextfieldPassword from "./components/fields/textfield_password.vue";
+import OnTextfieldText from "./components/fields/composed/textfield.vue";
+import OnTextfieldEmail from "./components/fields/composed/textfield_email.vue";
+import OnTextfieldNumber from "./components/fields/composed/textfield_number.vue";
+import OnTextfieldPassword from "./components/fields/composed/textfield_password.vue";
 import HiddenfieldHoneypot from "./components/fields/hiddenfield_honeypot.vue";
 import FormGridRow from "./components/containers/form_grid_row.vue";
 import OnTextArea from './components/fields/extended_textarea.vue'
@@ -20,21 +20,20 @@ import OnSelect from './components/fields/extended_select.vue'
 import SelectWithRelatedData from './components/fields/select_with_related_data.vue'
 import AdvancedPassword from "./components/fields/advanced_password.vue";
 import FileUpload from "./components/fields/file_upload.vue";
-import StaticText from "./components/fields/static_text/static_text.vue";
+import StaticText from "./components/fields/other/static_text/static_text.vue";
 import OnCaptcha from "./components/fields/onCaptcha/onCaptcha.vue";
 import DatePicker from "./components/fields/datepicker.vue";
-import MaskedText from "./components/fields/textfield_masked.vue";
+import MaskedText from "./components/fields/composed/textfield_masked.vue";
 import ConditionRadio from "./components/containers/condition_radio.vue";
 import ConditionCheckbox from "./components/containers/condition_checkbox.vue";
 import ConditionalContent from "./components/containers/conditional_content.vue";
-import Telephone from './components/fields/textfield_telephone.vue';
-import Url from './components/fields/textfield_url.vue';
+import Telephone from './components/fields/composed/textfield_telephone.vue';
+import Url from './components/fields/composed/textfield_url.vue';
 import MultiSelect from './components/fields/extended_multiselect.vue';
 import MultiCheckbox from './components/fields/multi_checkbox.vue';
-import ContentElement from './components/fields/content_element/content_element.vue';
+import ContentElement from './components/fields/other/content_element/content_element.vue';
 
 // INTERNAL ELEMENTS
-import InternalTextField from './components/fields/base/textfield/textfield.vue';
 import DynamicElement from './components/dynamic_element.vue';
 
 // PUBLIC OPTIONAL FUNCTIONS AND HELPERS
@@ -79,8 +78,7 @@ export {
     MultiSelect,
     MultiCheckbox,
     DatePicker,
-    ContentElement,
-    InternalTextField,
+    ContentElement
 }
 
 export { default } from '@/store/response_interceptor';