+++ src/stories/composed/TextFieldUrl.stories.js
@@ -1,0 +1,31 @@
+import {VTextField} from "vuetify/lib/components";
+import OnTextFieldUrl from '@/components/fields/composed/textfield_url'
+import Vue from "vue";
+import Vuetify from 'vuetify/lib/framework';
+import {defaultProps} from "@/stories/helper";
+
+Vue.use(Vuetify);
+const vuetify = new Vuetify({});
+
+Vue.component('v-text-field', VTextField)
+
+export default {
+    title: 'Formvue/Composed/TextFieldUrl',
+    component: OnTextFieldUrl,
+};
+
+const Template = (args, {argTypes}) => ({
+    props: Object.keys(argTypes),
+    vuetify: vuetify,
+    components: {TextField: OnTextFieldUrl},
+    template: '<text-field v-bind="$props" />',
+});
+
+export const Default = Template.bind({});
+Default.args = defaultProps('Url Input', 'url-input', false, false);
+Default.storyName = 'Default (optional)'
+
+export const Required = Template.bind({});
+// More on args: https://storybook.js.org/docs/vue/writing-stories/args
+Required.args = defaultProps('Url Input', 'url-input', false, true);
+