+++ changelog.md
@@ -8,6 +8,8 @@ import ResponseInterceptor from "formvue-json";
 ResponseInterceptor.register(async (ctx, jsonResponse) => false)
 ```
 
+- removed InternalTextField, as it was just TextField with no extra properties
+
 # 1.12.0
 - Removed 'menu' two-way-bind in extended_select and select_with_related_values as it seems to be a bad practise.
   Use '.activateMenu()', '.blur()' and '.isMenuActive' instead